Starlark transitions: better support list-typed options, stop crashes.
This is by no means a complete solution. We have a real challenge here in
that list-typed native options traditionally (before Starlark configuation)
can only take strings at the command-line. So their converters expect
inputs to be raw strings.

Now with SBC the values can be actual lists of Starlark objects. Those
lists may or may not type-match the option type. Because of the "string inputs
only" expectations, there's no easy way for the converters to check this.

Before this change, this could produce crashes (see GitHub bug) because
Blaze saw this as a list-typed option assigned to a list value, but didn't
check that the element type doesn't match (expected Label, actual string).

This change improves the situation, but only partially, and not for
allowMultiple options. See new code comments for details.

Fixes https://github.com/bazelbuild/bazel/issues/8936.

PiperOrigin-RevId: 308284734