Make sandboxfs work when the sandbox_base is not under .../tmp.
The sandbox-exec configuration we write covers the execroot of the sandbox
tree but not the sibling sandboxfs tree.  However, the configuration also
grants blanket write access to tmp directories.  As a result, when we use
the default output_base configuration, which puts it under tmp on macOS,
we get implicit write access to the sandboxfs tree -- but things break
when the output_base is moved elsewhere.

Fix this by adding the path to the execroot directory under the sandboxfs
tree to the sandbox-exec configuration.

Tested: This is impossible to test for automatically.  We'd need to find
a directory to which we can write from inside the test that is not covered
by the test's own sandbox.  Which means using a tmp directory, which
wouldn't work because they are writable, or disabling the sandbox.  But
if we disable the sandbox, we'd need to make assumptions about other
writable directories that are not tmp (e.g. the home directory), and using
those to put temporaries in is just wrong.  So... I tested this by hand
by building abseil with sandboxfs enabled and an output base under my
home directory (original repro in the bug).

Fixes #7435.

RELNOTES: None.
PiperOrigin-RevId: 308340444