bazel syntax: clean up resolution of functions
FunctionSignature (moved to Resolver.Function) is now
computed during resolution, not parsing. Its construction
piggybacks on the existing loops and hash tables used by
the Resolver's DefStatement case. It represent static
information about a function.

A Resolver.Function is annotated onto the syntax node for
each DefStatment, for the defined function, and onto StarlarkFile,
for the top-level function; also, resolveExpr returns a Function.

The design follows that of the go.starlark.net reference implementation:
https://github.com/google/starlark-go\
 /blob/master/resolve/resolve.go#L808

The Evaluator now requires that syntax trees have been successfully
resolved prior to evaluation, which is at long last true thanks
to CL 308109278 to Copybara.

I have relented in my earlier objections to execAndEvalOptionalFinalExpression,
and merged it with exec: that is, exec(StarlarkFile) returns the value
of the final expression statement, if any.

I have resisted the urge to improve the parameter logic
in this change, but have left TODOs for a follow-up.

See also:
- unknown commit, which added validation to copybara
- unknown commit, which added validation to skydoc

PiperOrigin-RevId: 308709027