Automated rollback of commit 77502091f06b84054921f4eeee3f1d093dd8427d.
*** Reason for rollback ***

Breaks some Bazel downstream projects.

*** Original change description ***

Remove calls to Artifact.getRunfilesPath() where the root-relative path is meant and rename the method to getShortPath() and add a scary warning.

getRunfilesPath() is ambiguous because artifacts in external repositories are available under two paths: external/$REPO/<path> and ../$REPO/<path>. The desired end state is for the former to go away, but it's a long road there.

RELNOTES: None.
PiperOrigin-RevId: 309003956