bazel syntax: don't reorder regular parameters
Prior to this change, function parameters were reordered
in resolveFunction so that required keyword-only parameters
precede optional ones, as the source permits them to be mixed.

This change keeps all parameters (except *args) in source order,
and instead uses a sentinel value, MANDATORY, in the
StarlarkFunction.defaultValues tuple, to indicate the positions
of mandatory parameters, slightly simplifying the algorithms
for argument processing.

Importantly, parameter Identifiers are now bound in run-time
order, which sets us up for using the function's binding array
as the list of parameter names.

All of this is a step to convergence with the go.starlark.net resolver.

Also:
- issue more specific errors during parameter validation,
  and remove partial validation from parser.
- rename Resolver.addError -> errorf and use String.format
- rename Resolver.declare -> bind
- add missing tests of parameter validation.
PiperOrigin-RevId: 309046975