bazel syntax: introduce Resolver.Binding
This refactoring is a first step towards a go.starlark.net-style
one-pass resolver that produces flat environment frames.

A Binding is a static abstraction of a variable.
The Resolver's main job is to resolve each Identifier
to its Binding. Bindings are created by binding
occurrences of a name: loads, parameters, assignments.

Also:
- cleanups to Scope enum
- remove specialized errors for long-obsolete {PACKAGE,REPOSITORY}_NAME.
PiperOrigin-RevId: 309080798