Remove unnecessary order dependency between direct/transitive children nodes of ArtifactNestedSetKey.
ArtifactNestedSetFunction makes 2 getValues calls for direct and transitive
children. This results in an implicit ordering, and 2 different dependency
groups, when evaluating the nodes. All this amounts to an increased wall
time in incremental builds since these deps are unnecessarily evaluated sequentially. We've therefore combined these 2 calls into 1 getValuesOrThrow. env.getValuesOrThrow because we need to access, in ActionExecutionFunction#accumulateInputs[1], the Exceptions thrown while evaluating the dependencies.

However, using env.getValuesOrThrow means that the SkyFunction is in charge of handling the Exceptions from the evaluation. This CL adds that to ArtifactNestedSetFunction, with the mandate from SkyFunction[2].

This CL also changes the condition in which an Artifact makes it to ArtifactNestedSetFunction#artifactSkyKeyToValueOrException: only when a non-null value or exception is present. There can be evaluations in ArtifactNestedSetFunction where not all the dependencies are present (i.e. env.valuesMissing() == true), but some of them already throw an error and we need to stop the evaluation to bubble it up. ActionExecutionFunction can then retrieve the Artifact's exception via ArtifactNestedSetFunction#artifactSkyKeyToValueOrException without null values tripping it.

[1]https://source.bazel.build/bazel/+/master:src/main/java/com/google/devtools/build/lib/skyframe/ActionExecutionFunction.java;l=1171;drc=b20f5122b9523381d70b52ebbb41a9a22e96b87b
[2]https://source.bazel.build/bazel/+/master:src/main/java/com/google/devtools/build/skyframe/SkyFunction.java;l=253-256;drc=b20f5122b9523381d70b52ebbb41a9a22e96b87b

RELNOTES: None
PiperOrigin-RevId: 309237438