Remove dead code for `mapAll` from `Args.add_all`.
The implementation of `Args.add_all` has a flag for a `map_all` parameter with
code supporting to evaluate and construct keys based on that, however it is
always set to null, deeming all of the logic a no-op. Historically, it was used
to implement `map_fn` for `Args.add`, but this has been dropped already.

Delete the unused `map_all` parameter and code used to support it.

PiperOrigin-RevId: 309475789