Properly handle package-default visibility in AggregatingAttributeMapper
Have `#visitAttribute` recognize that visibility is special, which lets it play
nice with the `attr` query function and lets us delete
`#getPossibleAttributeValues`. We check for the visibility special case after
checking for selects and computed defaults - visibility shouldn't be either of
these, but stranger things have happened.

I think the history here is that package-default visibility was always broken
for the `attr` query function, which uses `#visitAttribute`, while query
output was correct because it used #getPossibleAttributeValues (or other
scattered special-casing which existed at various times).

rule_test_test needed to be updated because it was expecting the broken
behavior - package-default visibility being ignored by `attr(...)`.

PiperOrigin-RevId: 309477218