Have --incompatible_string_replace_count reject None counts
This is another change to bring string.replace() in line with the spec. Safe to modify the flag since it has not been released yet.

Eliminated use of parameterization in test suite since it caused test cases that didn't need it to be cloned.

Also add RELNOTES in this CL for the flag.

Follow-up work to #9181.

RELNOTES: Added --incompatible_string_replace_count, to make string.replace() behave as in Python.
PiperOrigin-RevId: 309483323