Provide a way to build against Debian system installed java libraries
This change provides a way to replace some java dependencies in third_party with Debian system installed java libraries.

- Add a new local repository which points to /usr/share/java
- Define targets for needed system java libraries in `debian_java.BUILD"
- Add a new config_setting to enable the Debian build
- Introduced a new distrib_java_import rule to help redirect targets to the ones defined in debian_java.BUILD
With the above work, we can replace a jar dependency by two steps:

- Define the target in debian_java.BUILD
- Add enable_distributions = ["debian"] attribute for the target we want to replace.

To enable building against Debian system libs, add --define=distribution=debian as a build option when building Bazel.

To add a new distribution support for Bazel, we only need to

- Add a new config_setting definition for the distribution
- Implement the logic for redirecting targets for the distribution in tools/distributions/distribution_rules.bzl
- Add the distribution name to enable_distributions attribute for the targets we want to replace.

Note that this is for Bazel's java dependencies only, we'll implement the similar procedure for our C++ dependencies in later changes.

Related #9408

Closes #11282.

PiperOrigin-RevId: 309781262