Add the OS as a field to the configuration
This helps move Bazel towards being able to perform cross-platform builds
with a remote execution system that's running a different OS. This is in
preparation for updating the rule implementations to use the configuration
setting instead of OS.current() directly.

I'm not sure how this interacts with platforms, but this change seems safe
enough.

Progress on #11748.

Change-Id: I20dff1fb332d4567b621d65e43b05a00905f75b0