Fix no-remote-exec caching functionality
This commit addresses issue #9412 and provides all
instances of the RemoteSpawnStrategy with a RemoteSpawnCache
to ensure that the AbstractSpawnStrategy caching logic behaves
as expected for no-remote-exec actions.

This patch was written for 1.2.1 and adapted to latest master.

EDIT: Force pushed different changes 19/03/2020, see below for further discussion
EDIT: Rebased as of 28/04/20, unsure if 08efd761721 should be needed

Closes #10964.

PiperOrigin-RevId: 309943659