Replace "" (empty string) default values of multiple options with "null" (round 2).
This is preparation for turning on parsing of default values for multiple (allowMultiple = true) options. This change is NOOP (enforced by OptionDefaultValueConversionTest#shouldConvertDefaultValue()) because for now default value of a multiple option does not matter and is always converted to an empty list. But it will matter a bit later: "null" - special value resolving to the same empty list, anything else - parsed by the converter.

RELNOTES: None.
PiperOrigin-RevId: 310577360