bazel syntax: replace map with callback for module loading
This change replaces the map of Modules, keyed by name, provided
at thread creation, with a callback, so that clients can implement
loading lazily. The interpreter has no need to enumerate modules.

Also, document next steps for long-awaited decoupling of thread
and module.

This is a breaking API change for Copybara.

PiperOrigin-RevId: 310614969