Add --compilation_mode support in transitions
Add support for converting //command_line_option:compilation_mode from
Java value to Starlark value so that it can be read in a transition.

Implementing StarlarkValue is a workaround until a well-defined
Java-Starlark conversion interface has been created. See also
https://github.com/bazelbuild/bazel/pull/11347#issuecomment-630260102
Avoid replicating this workaround.

Fixes --compilation_mode in #10690, but not the general case.