Implement stricter checks for adding shared library as non-versioned library.
A library provided with the name "foo.so" must be linked to using
"-l:foo.so" and not "-lfoo" (which would map to "libfoo.so") using the
"versioned shared library" logic (which correctly handles libraries with
names like "libfoo.so.1a.2").

This change adds further checks so only libraries with names like
"libX.so" are treated as non-versioned libraries on the link command
line.

Closes #10826.

PiperOrigin-RevId: 310927514