Fold killDelay into ProcessWrapper.
The kill delay setting comes from the --local_termination_grace_seconds
option (it's not specific to individual process-wrapper invocations) and
we want every process-wrapper invocation to respect it.

In the past, we were propagating this value through various layers so
that we could reach the places where command lines were built... but we
can now take advantage of the ProcessWrapper object to pass this
information around.

Part of https://github.com/bazelbuild/bazel/issues/10245.

RELNOTES: None.
PiperOrigin-RevId: 311112402