Provide transitive digest of bzl files for each Starlark Module.
Replace the optional label in Module with a context object, storing arbitrary
additional data. The context will still carry a label which can be used to
enhance the information provided by the repr, but also allows to store more
information, useful outside of Starlark interpreter. An example of this is the
transitive hash of the bzl files loaded by the module, which is not relevant to
Starlark itself, but is important in Bazel.

Move the concept of bzlTransitiveDigest from BazelStarlakContext (associated
with the thread) to new BazelModuleContext, which is added to the Starlark
module.

The concept of the transitive digest makes most sense at the level of the
module -- it means the `digest({bzl \in transitively_loaded(module)})`. After
this change, we rely on picking appropriate module to get the digest from
rather than the data associated with the StarlarkThread.

One of the uses of the bzlTransitiveDigest is adding a digest for RuleClass
instances. After this change, we explicitly use the hash of the module which
defines the RuleClass itself.

PiperOrigin-RevId: 311173293