Log top package loading events
Useful for understanding where time goes in the loading phase without needing to
explicitly use a profiler.

The logging should only happen when packages are loaded - no need to be
selective about what commands this kicks in for.

I went back and forth on being able to toggle the number of extrema to track,
wound up throwing my hands up and deciding to leave it hard coded for now. We
can bound it in another change (gets tricky with how BlazeModules work).

PiperOrigin-RevId: 311205729