Remove treeArtifactContents from OutputStore.
treeArtifactContents acts as a temporary cache prior to the full tree artifact being constructed and is optional - if not populated, ActionMetadataHandler reads tree artifact contents from disk. In fact, even if it is populated, disk reads are performed to verify expected contents.

Callers can use injectRemoteDirectory to establish known tree artifact contents instead. In this case, we assume that the contents are correct, so there is no need to check for a match on disk.

PiperOrigin-RevId: 311241850