Make the number of packages we log metrics for configurable and add tests
Sets the default to 10, down from the original 50, which could generate a lot
of output.

This cl is mostly boilerplate in tests. On one hand I may have gone overboard,
on the other this makes it hard to avoid adding tests if functionality becomes
more complex, and I'd feel awfully bad if a silly mistake here caused a crash
bug.

PiperOrigin-RevId: 311247814