bazel syntax: start disentangling EvaluationTestCase
This change causes various tests that formerly extended
SkylarkTestCase to extend its superclass, BuildViewTestCase.
Inheritance is a poor means of code re-use as it causes
things to become entangled. For example construction of
the test helper is coupled to construction of the Junit
test suite.

Now, those tests embed an EvaluationTestCase (really an
"apparatus" in Blaze testing jargon) as a field and refer
to it explicitly.

SkylarkTestCase's anonymous subclass of EvaluationTestCase,
which predeclares the Bazel built-ins, has been named as
BuildEvaluationTestCase. SkylarkTestCase is deleted.

Also, eliminate calls to reset() / initialize() now
that setSemantics reinitializes the thread.

The test structure is still a mess, but this change unblocks
a change to the production API that I have been trying
in vain to submit for weeks.

I have lost count of the number of good CLs I have had
to abandon because the entanglement caused by
SkylarkTestCase made updating the tests impossible.
Good riddance.

PiperOrigin-RevId: 311252614