Fix crash when config_setting values are parsed as multiple values.
config_setting's contract says that for flag_values = {"key": "val"},
"val" must be a single value that "key" can take. This gets confusing
for multi-value options, where if we pass {"key": "a,b"} the option
converter may try to turn this into ["a", "b"], which would violate
that contract.

Note that ["a,b"] would be perfectly fine (that's a single value in
the list that just happens to have commas in it). It really depends
on the specific option's converter what happens.

Fixes https://github.com/bazelbuild/bazel/issues/10146

PiperOrigin-RevId: 311325708