Move documentation of 'visibility' to a separate page
Visibility is a core concept (you have to know about when you create a 2nd package).
It shouldn't be hidden in a table, inside the list of common attributes.

I also plan to expand the documentation. Creating a proper section for visibility will leave more space for structuring the content.

This change doesn't modify the content itself (except for one introduction sentence).

https://github.com/bazelbuild/bazel/issues/8982

RELNOTES: None.
PiperOrigin-RevId: 311371192