Fix handling of default licenses and distribs in query
Manual roll-forward of https://github.com/bazelbuild/bazel/commit/d5f918f2a2574f7363c9ef1ad9c01c097b447b95 with fixes for query's `attrs()`s handling
of licenses and distribs adapted from https://github.com/bazelbuild/bazel/commit/0f38eead54a598f3489ad877146cafacff26b3ab.

To recap...

Rule.getLicense() and Rule.getAttribute("license") could fall out of sync if
the package-default value was set via `licenses` or `pacakages`. Same goes for
distribs. This is confusing and wrong for users using query to understand
licenses and distribs. This could also lead to the `attr()` query function
seeing incorrect or misleading values for licenses and distribs.

The fix is to also special-case licenses and distribs in
AggregatingAttributeMapper#visitAttributes. This _should_ be a "temporary" fix,
as licenses and distribs are going away.

Note that this does uncover another awkwardness with licenses and distribs -
`attr()` sees distribs' values in all UPPER_CASE, while it sees licenses'
values in lower_case. Given this is all allegedly temporary, I'm leaving this
as-is as WAI for now.

Note that applicable_licenses suffers from the same problems, but it's broken
in more ways, so ignoring it for now.

PiperOrigin-RevId: 311389441