bazel syntax: stop creating wall-time profiler spans for each function call
We have a CPU profiler now, which is more accurate in the presence of
parallelism, as every runnable but not running thread accumulates wall
time even as it does nothing. (Also, I saw no sign of these spans in
when running blaze query --experimental_generate_json_trace_profile deps(X)
where X is a Substantial Program.)

The higher-level unit of BUILD/.bzl file execution is a more appropriate span.

PiperOrigin-RevId: 311411836