Automated rollback of commit 5d98324022a156542b768849d4dedcde49eb67d0.
*** Reason for rollback ***

This was helpful in the past and recently added on request of Bazel users.

*** Original change description ***

bazel syntax: stop creating wall-time profiler spans for each function call

We have a CPU profiler now, which is more accurate in the presence of
parallelism, as every runnable but not running thread accumulates wall
time even as it does nothing. (Also, I saw no sign of these spans in
when running blaze query --experimental_generate_json_trace_profile deps(X)
where X is a Substantial Program.)

The higher-level unit of BUILD/.bzl file execution is a more appropriate span.

PiperOrigin-RevId: 311514345