java.starlark.net: rename com.google.devtools.starlark
This change renames the Java packages of the Starlark
interpreter and REPL command.

The Bazel team owns the domain starlark.net, and
the Go implementation lives at go.starlark.net;
the Java packages should use java.starlark.net.

There are no immediate plans to publish the interpreter
as a public API or to run a web server at that domain,
but it will save pain later if we rename the classes only once.

Now that the Starlark interpreter's production code
no longer depends on any Bazel packages (except serialization,
which cannot be broken until we implement byte compilation),
we will soon make the following moves:

  lib.syntax (frontend)  -> net.starlark.java.syntax
  lib.syntax (evaluator) -> net.starlark.java.eval
  lib.skylarkinterface   -> net.starlark.java.annot

PiperOrigin-RevId: 312126212