Allow setting test_tmpdir to outside of execroot for `blaze run`.
When running a test target using `blaze run`, we can specify the
`--test_tmpdir` parameter, however trying to do so with a folder outside of the
exec root results with a failure.

The root cause of the issue is that we assume that `TEST_TMP_DIR` is under the
exec root and try to get a relative path within it. StandaloneTestStrategy, for
example, has a check to only do so if `TEST_TMP_DIR` is underneath the exec
root.

Add a similar check to only resolve the `TEST_TMP_DIR` as relative to the exec
root if it truly is underneath it.

PiperOrigin-RevId: 312152035