Reenable fsevents tests.
I've fixed a bunch of additional problems in fsevents (incorrect
handling of directory notifications and spurious empty paths) and the
tests (inaccurate change detection) during the last few days.

I do not know if the original flakiness we observed in CI was caused
by any of these, but I just cannot get the tests to flake in the same
way by hand -- not even in one of the CI machines... so let's give
them another try.

Fixes #10776 (or so I hope).

RELNOTES: None.
PiperOrigin-RevId: 312282170