Fix windows --watchfs
As discussed in #1931 --watchfs was made a no-op on windows in d04d7baa1605c8c3a637255bd61c00dba73ad800. The reason for this is that it will hold locks for the (parent-directories of) directories it is watching. Since we recursively register the watcher this leads to locks on folders we do not want to be locked.

The issue is known and tracked at https://bugs.openjdk.java.net/browse/JDK-6972833 but is marked as a Won't Fix. The suggested fix is to instead use ExtendedWatchEventModifier.FILE_TREE which allows a single WatchKey to recursively watch a folder on windows.

This pull request implements the FILE_TREE flag for the root folder that is being watched, and disables the registration of the recursive folders. I have tried to keep the change as small as possible, but I can imagine that we wish to separate this logic to dedicated windows class similar to the OSX variant.

Fixes #1931

Closes #11334.

PiperOrigin-RevId: 312285715