Add override for worker mnemonic via execution requirements
Sometimes the same (heavy) binary is used for different actions that may
have different performance characteristics. Without a way to separate the
pooling behavior from the reporting data, rule maintainers have to choose
between fine-grained reporting on the different kinds of actions and
resource usage due to unnecessary worker processes.

RELNOTES: It is now possible to use different action mnemonics while still
sharing the same pool of persistent workers. It requires setting a new property
on the execution requirements (`worker-key-mnemonic`). The value overrides
the action's mnemonic when it comes to reusing worker processes.
PiperOrigin-RevId: 312490508