Prune description of InstrumentedFilesInfo
The bit about InstrumentedFilesInfo signaling that a rule provides _source_ files that could be instrumented is a bit misleading. InstrumentedFilesInfo should also be configured if a rule provides _dependencies_ that could be instrumented. That is, any rule that might provide (directly or indirectly) code or binaries that might have code coverage instrumentation and might be executed at runtime should configure InstrumentedFilesInfo. This is basically every rule.

RELNOTES: None.
PiperOrigin-RevId: 312583400