Rename ActionExecutionValue#createSimpleFileArtifactValue and make it an instance method.
The name of the method is misleading because it doesn't create anything new, just gets it from a map. It also asserts that it exists, so rename it to getExistingFileArtifactValue.

Also, make it an instance method. A static method that requires passing in an instance of the class seems awkward.

RELNOTES: None.
PiperOrigin-RevId: 312712401