Tighten the types of some of the machinery surrounding .bzl skyframe inlining
Tighten SkyKey to StarlarkImportLookupValue.Key for the internal functions of StarlarkImportLookupFunction and its inlined (non-Skyframe) caller. This avoids all but one downcast to [...].Key, in compute(), as is expected. Note the naming convention for vars -- "SkyKey skyKey", but "StarlarkImportLookupFunction.Key key".

Also eliminate a redundant case in PackageFunction#computeStarlarkImportMapWithInlining, at the expense of making the quick return path not self-evident; added a comment to compensate.

This is refactoring work toward adding a new kind of .bzl loading context, for Bazel-internal .bzl files.

Work toward #11437.

RELNOTES: None
PiperOrigin-RevId: 312840745