Don't split up "Writing file" string.
In order to discover the source of the message "Writing file $path"
I did a code search for the string '\"Writing\ file' but this did
not find anything useful. In the end I found the source of this
message but it would have been easier to search for if the
string was not split up.

PiperOrigin-RevId: 313253529