Add a new option --experimental_nsos_eval_keys_as_one_group.
The flag's purpose is to enable the evaluation of the children of each ArtifactNestedSetKey as a single unit instead of in 2 separate groups: direct and transitive children.

The flag is unused in this CL and the logic will be added in a follow-up CL.

Like the NestedSet size threshold, the evaluator is reset if this option
changes between runs to maintain Skyframe dep consistency.

RELNOTES: None
PiperOrigin-RevId: 313385537