Move the addition of SpawnMetrics to its Builder
This is a refactoring that will make it a bit easier to reuse the code to add
metrics. This should be useful when we to aggregate the metrics per different
execution mode (remote/local/worker).

Note that this actually does change the behavior in a small way. Previously
calling `aggregateMetrics(..., /* onlyDuration= */ true)` would result in
setting the non-duration fields to the value contained in the last `SpawnMetrics`
in the list. Now, we'll take the maximum value for each of them.

RELNOTES: None.
PiperOrigin-RevId: 313387263